refactor: All the items of the same type will show up
Every shoe Id is in this form:
Shoe model + Male or Female + Kid or Adult + Color.
We will show all the shoes that are the same but different colors.
For the accessories the format is different:
Item type + color + size.
There will be the items with the same type and color but different sizes.